Securisation access to any Controllers & twig
Securisation AuteurController, EmpruntController,GenreController, LivreController
Securisation templates: emprunt, fron, livre, base.html.twig

+++ src/Controller/LivreController.php
@@ -9,6 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('admin/livre')]
 class LivreController extends AbstractController
@@ -16,6 +17,8 @@ class LivreController extends AbstractController
     #[Route('/', name: 'app_livre_index', methods: ['GET'])]
     public function index(LivreRepository $livreRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
+
         return $this->render('livre/index.html.twig', [
             'livres' => $livreRepository->findAll(),
         ]);
@@ -24,6 +27,8 @@ class LivreController extends AbstractController
     #[Route('/new', name: 'app_livre_new', methods: ['GET', 'POST'])]
     public function new(Request $request, LivreRepository $livreRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
+
         $livre = new Livre();
         $form = $this->createForm(LivreType::class, $livre);
         $form->handleRequest($request);
@@ -43,14 +48,26 @@ class LivreController extends AbstractController
     #[Route('/{id}', name: 'app_livre_show', methods: ['GET'])]
     public function show(Livre $livre): Response
     {
-        return $this->render('livre/show.html.twig', [
-            'livre' => $livre,
-        ]);
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('livre/show.html.twig', [
+                'livre' => $livre,
+            ]);;
+
+        } elseif ($this->isGranted('ROLE_EMPRUNTEUR')) {
+            return $this->render('front/livre.html.twig', [
+                'livre' => $livre,
+            ]);
+        }
+        
     }
 
     #[Route('/{id}/edit', name: 'app_livre_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Livre $livre, LivreRepository $livreRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
+
         $form = $this->createForm(LivreType::class, $livre);
         $form->handleRequest($request);
 
@@ -69,6 +86,8 @@ class LivreController extends AbstractController
     #[Route('/{id}', name: 'app_livre_delete', methods: ['POST'])]
     public function delete(Request $request, Livre $livre, LivreRepository $livreRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
+
         if ($this->isCsrfTokenValid('delete'.$livre->getId(), $request->request->get('_token'))) {
             $livreRepository->remove($livre, true);
         }
